DRY'd up code for updating task list container
