+++ public/javascripts/fetch-tasks.js
@@ -1,4 +1,4 @@
-import { addTaskInfoListeners, updateTotalTaskValue, updateOverDueValue, updateTasksCompletedValue } from "./utils.js";
+import { addTaskInfoListeners, updateTotalTaskValue, updateOverDueValue, updateTasksCompletedValue, updateTaskListContainer } from "./utils.js";
 
 // fetch user's tasks (all)
 export const fetchTasks = async () => {
@@ -11,27 +11,12 @@ export const fetchTasks = async () => {
 
   const { tasks, user } = await res.json();
 
-  const tasksListContainer = document.querySelector(".task-list");
   const listName = `
   <h2 class="task-list-header"><strong>All of My Tasks</strong></h2>
   `
-  const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-    if (isCompleted === true) {
-      return `
-        <div class='task-info' id=${id}>
-            <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-            <label for=${id} id=${id} class="task-check-box">${description}</label>
-        </div>
-        `
-    } else {
-      return `<div class='task-info' id=${id}>
-                <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-                <label for=${id} id=${id} class="task-check-box">${description}</label>
-            </div>
-        `
-    }
-  })
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
+
+  await updateTaskListContainer(tasks, listName);
+
   await addTaskInfoListeners();
   await updateOverDueValue();
   await updateTotalTaskValue();
@@ -86,28 +71,10 @@ export const fetchIncompleteTasks = async () => {
 
   const { tasks, user } = await res.json();
 
-  const tasksListContainer = document.querySelector(".task-list");
   const listName = `
     <h2 class="task-list-header"><strong>Still Need To Brew</strong></h2>
     `
-  const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-    if (isCompleted === true) {
-      return `
-        <div class='task-info' id=${id}>
-            <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-            <label for=${id} id=${id} class="task-check-box">${description}</label>
-        </div>
-        `
-    } else {
-      return `<div class='task-info' id=${id}>
-                <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-                <label for=${id} id=${id} class="task-check-box">${description}</label>
-            </div>
-        `
-    }
-  })
-
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
+  await updateTaskListContainer(tasks, listName);
 
   await addTaskInfoListeners();
 }
@@ -124,28 +91,10 @@ export const fetchCompletedTasks = async () => {
 
   const { tasks, user } = await res.json();
 
-  const tasksListContainer = document.querySelector(".task-list");
   const listName = `
     <h2 class="task-list-header"><strong>Got a latte done!</strong></h2>
     `
-  const tasksHtml = tasks.map(({ id, description, isCompleted }) => {
-    if (isCompleted === true) {
-      return `
-        <div class='task-info' id=${id}>
-            <input type="checkbox" class="task-check-box" id=${id} name=${id} checked>
-            <label for=${id} id=${id} class="task-check-box">${description}</label>
-        </div>
-        `
-    } else {
-      return `<div class='task-info' id=${id}>
-                <input type="checkbox" class="task-check-box" id=${id} name=${id}>
-                <label for=${id} id=${id} class="task-check-box">${description}</label>
-            </div>
-        `
-    }
-  })
-
-  tasksListContainer.innerHTML = listName + tasksHtml.join("");
+  await updateTaskListContainer(tasks, listName);
 
   await addTaskInfoListeners();
 }
